modified some code to make it more readable and depending on user-input
