+++ ITI-tasks/task16.py
