+++ ITI-tasks/task18/task18.py
