+++ ITI-tasks/task2.py
@@ -1,8 +1,12 @@
-#!/bin/pyhon3
+#!/bin/python3
 
-# 2- Write a Python program to sum all the items in a list. --> [2,7,10,18]
+# 2- Write a Python program to sum all the items in a list.
 
-myFirstList = [2,7,10,18]
+# Take the list from the user
+user_input = input("Enter a list of numbers separated by spaces: ")
+
+# Convert the input string into a list of integers
+myFirstList = [int(x) for x in user_input.split()]
 
 listSum = 0
