+++ ITI-tasks/task22/task22.py
