+++ ITI-tasks/task23/task23.py
