+++ ITI-tasks/task8.py
@@ -2,14 +2,25 @@
 
 # Write a Python program to sum all the items in a dictionary.
 
-myDict = {'data1': 100, 'data2': -54, 'data3': 247}
+# Function to get a dictionary from the user
+def get_user_dict():
+    user_dict = {}
+    print("Enter key-value pairs for the dictionary (type 'done' to finish):")
+    while True:
+        key = input("Enter key (or 'done' to stop): ")
+        if key.lower() == 'done':
+            break
+        try:
+            value = int(input(f"Enter value for key '{key}': "))
+            user_dict[key] = value
+        except ValueError:
+            print("Invalid value. Please enter an integer.")
+    return user_dict
 
-sum = 0
+# Get the dictionary from the user
+myDict = get_user_dict()
 
-for key in myDict:
+# Calculate the sum of all values in the dictionary
+total_sum = sum(myDict.values())
 
-    sum += myDict[key]
-
-
-print(sum)
-
+print("The sum of all items in the dictionary is:", total_sum)
